Accumulated Local Effects for ERF

Calculates the Accumulated Local Effects (ALE) from an ERF object

ALE_fn(
  fit,
  var,
  save = TRUE,
  out.folder = NULL,
  cores = parallel::detectCores() - 4
)

Arguments

fit
The fitted object returned from calling ens_random_forests()

var
The name of the response variable

save
A logical flag to save the output as an RData object, default is TRUE.

out.folder
A path to the folder to write out too. If NULL then a folder is generated in the working directory

cores
An integer value that either indicates the number of cores to use for parallel processing or a negative value to indicate the number of cores to leave free. Default is to leave two cores free.

Value

A list that contains a data.frame for each variable, ordered by the mean variable importance, and a vector of the covariate values (used for rug plot in plot_ALE). The columns in each data.frame are as follows:

x: the covariate values that the ALE was calculated for

class: the class of the covariate; used by subsequent plot_ALE function

q: the quantile of the x value of the covariate

f.X: the ALEs evaluated at a given x value

Examples

#run an ERF with 10 RFs and 
ens_rf_ex <- ens_random_forests(df=simData$samples, var="obs", covariates=grep("cov", colnames(simData$samples),value=T), save=FALSE, cores=1)

#> rounding n.forests to the nearest one

ALEdf <- ALE_fn(ens_rf_ex, save=FALSE)

#> No name of response variable, making one

#> rounding n.forests to the nearest one

head(ALEdf[[1]]$df)

#>            x   class      q        f.1        f.2         f.3        f.4
#> 1 -0.6321187 numeric 0.0002 -0.1226546 -0.1790335 -0.06127724 -0.1165853
#> 2 -0.5269505 numeric 0.0200 -0.2376546 -0.1790335 -0.20627724 -0.2065853
#> 3 -0.4847148 numeric 0.0401 -0.2376546 -0.1790335 -0.20627724 -0.2065853
#> 4 -0.4507819 numeric 0.0601 -0.2276546 -0.1740335 -0.21627724 -0.1965853
#> 5 -0.4220293 numeric 0.0801 -0.2176546 -0.1640335 -0.20627724 -0.1865853
#> 6 -0.3994806 numeric 0.1001 -0.2176546 -0.1640335 -0.22127724 -0.1915853
#>           f.5         f.6        f.7         f.8        f.9        f.10
#> 1 -0.04644132 -0.09909558 -0.1177094 -0.06508035 -0.1345466 -0.07546232
#> 2 -0.20644132 -0.16909558 -0.2627094 -0.21008035 -0.2195466 -0.20546232
#> 3 -0.20644132 -0.16909558 -0.2627094 -0.20510522 -0.2195466 -0.20546232
#> 4 -0.20644132 -0.16909558 -0.2527094 -0.20010522 -0.1995466 -0.19046232
#> 5 -0.20144132 -0.16409558 -0.2477094 -0.17510522 -0.1945466 -0.17046232
#> 6 -0.21144132 -0.16909558 -0.2477094 -0.17510522 -0.2145466 -0.18546232