Accumulated Local Effects for ERF

Calculates the Accumulated Local Effects (ALE) from an ERF object

calc_ALE(
  fit,
  var,
  save = TRUE,
  out.folder = NULL,
  cores = parallel::detectCores() - 4
)

Arguments

fit
The fitted object returned from calling ens_random_forests()

var
The name of the response variable

save
A logical flag to save the output as an RData object, default is TRUE.

out.folder
A path to the folder to write out too. If NULL then a folder is generated in the working directory

cores
An integer value that either indicates the number of cores to use for parallel processing or a negative value to indicate the number of cores to leave free. Default is to leave two cores free.

Value

A list that contains a data.frame for each variable, ordered by the mean variable importance, and a vector of the covariate values (used for rug plot in plot_ALE). The columns in each data.frame are as follows:

x: the covariate values that the ALE was calculated for

class: the class of the covariate; used by subsequent plot_ALE function

q: the quantile of the x value of the covariate

f.X: the ALEs evaluated at a given x value

Examples

#run an ERF with 10 RFs and 
ens_rf_ex <- ens_random_forests(df=simData$samples, var="obs", covariates=grep("cov", colnames(simData$samples),value=T), save=FALSE, cores=1)

#> rounding n.forests to the nearest one

ALEdf <- calc_ALE(ens_rf_ex, save=FALSE)

#> No name of response variable, making one

#> rounding n.forests to the nearest one

head(ALEdf[[1]]$df)

#>            x   class      q        f.1        f.2         f.3        f.4
#> 1 -0.6321187 numeric 0.0002 -0.1124345 -0.1962214 -0.06231669 -0.1111357
#> 2 -0.5269505 numeric 0.0200 -0.2274345 -0.1962214 -0.20731669 -0.2011357
#> 3 -0.4847148 numeric 0.0401 -0.2274345 -0.1962214 -0.20731669 -0.2011357
#> 4 -0.4507819 numeric 0.0601 -0.2174345 -0.1912214 -0.21731669 -0.1911357
#> 5 -0.4220293 numeric 0.0801 -0.2074345 -0.1812214 -0.20731669 -0.1811357
#> 6 -0.3994806 numeric 0.1001 -0.2074345 -0.1812214 -0.22231669 -0.1911357
#>           f.5         f.6        f.7         f.8       f.9        f.10
#> 1 -0.04450836 -0.08992608 -0.1069719 -0.06323448 -0.142773 -0.07537429
#> 2 -0.20450836 -0.15992608 -0.2519719 -0.20823448 -0.227773 -0.20537429
#> 3 -0.20450836 -0.15992608 -0.2519719 -0.20325935 -0.227773 -0.20537429
#> 4 -0.20450836 -0.15992608 -0.2419719 -0.19825935 -0.207773 -0.19037429
#> 5 -0.19950836 -0.15492608 -0.2369719 -0.17825935 -0.202773 -0.17037429
#> 6 -0.20950836 -0.15992608 -0.2419719 -0.17825935 -0.222773 -0.18537429