Ensemble Random Forests

Execute Ensemble Random Forests model on a given dataset

ens_random_forests(
  df,
  var,
  covariates,
  header = NULL,
  out.folder = NULL,
  duplicate = TRUE,
  n.forests = 10,
  cores = parallel::detectCores() - 2,
  save = TRUE,
  ntree = 1000,
  mtry = 5
)

Arguments

df
A data.frame object

var
A character string indicating the column name of the data frame that contains the number of interactions for the ERF to model; column should be a numeric column

covariates
A character vector indicating the column name(s) of the data frame that contain the covariates

header
A character vector indicating the column name(s) of the data frame that contain the additional columns you wish appended to the output

out.folder
A path to the folder to write out too. If NULL then a folder is generated in the working directory

duplicate
A logical flag that indicates whether to duplicate observations with more than one interaction. Default is TRUE to duplicate all records that interacted with more than one individual (i.e. a fishing set that caught two of the same species)

n.forests
A numeric value indicating how many Random Forests to generate in the ensemble, default is 100

cores
A numeric value that either indicates the number of cores to use for parallel processing or a negative value to indicate the number of cores to leave free. Default is to leave two cores free.

save
A logical flag to save the output as an RData object, default is TRUE.

ntree
The number of decision trees to use in each RF, default is 1000

mtry
The number of covariates to try at each node split, default is 5

Value

A fitted ERF model as a list object. Contains the model data, the fitted RF models, the ensemble predictions, the ensemble model performance, the mean training/test performance, the ROC training/test performance for each RF, and the individual RF predictions.

Examples

#run an ERF with 10 RFs and 
ens_rf_ex <- ens_random_forests(df=simData$samples, var="obs", covariates=grep("cov",colnames(simData$samples),value=T), save=FALSE, cores=1)
head(ens_rf_ex$data) # view the dataset used in the model

#>   obs        cov1        cov2       cov3        cov4        cov5     random
#> 1   0 -0.07158000 -0.27811766  0.5741324  0.01366734 -0.15537696 -0.8543624
#> 2   0 -0.12028137  0.26103341 -0.2298234  0.14750795  0.32482174  0.6725291
#> 3   0  0.12422049  0.09052045 -0.3070963  0.14493653  0.02303010  1.0337192
#> 4   0 -0.12481261 -0.29583389  0.6282519  0.06067430 -0.25240678  0.4702388
#> 5   0  0.01361460 -0.04989738  0.1600218  0.42189437  0.13166581 -1.4174715
#> 6   0  0.01842766  0.18300638 -0.3421439 -0.33128857  0.07998399  0.9871635

head(ens_rf_ex$ens.pred) #view the model predictions

#>      P.0    P.1 PRES   resid
#> 1 0.7247 0.2753    0 -0.2753
#> 2 0.3872 0.6128    0 -0.6128
#> 3 0.6012 0.3988    0 -0.3988
#> 4 0.7456 0.2544    0 -0.2544
#> 5 0.3286 0.6714    0 -0.6714
#> 6 0.9910 0.0090    0 -0.0090

ens_rf_ex$mu.te.perf #view the mean test threshold-free performance metrics

#>     teAUC    teRMSE     teTSS 
#> 0.7901946 0.3904004 0.5299979 

unlist(ens_rf_ex$ens.perf[c('auc','rmse','tss')]) #view the threshold-free ensemble performance metrics

#>       auc      rmse       tss 
#> 0.9672889 0.3816819 0.8044123