Determines the maximum sample size in the second RF bag

Takes a data.frame object determines the maximum split

max_splitter(v, p = 0.89, nmax = 10000)

Arguments

v

A data frame object returned from erf_data_prep()

p

A numeric value between (0,1), default is just below 90%

nmax

A numeric value > 0 specifying the maximum number of observations per bag, default is 1e4

Value

A numeric value specifying the maximum split

Examples

data <- erf_data_prep(df=simData$samples, var='obs', covariates=grep('cov', colnames(simData$samples), value=TRUE))
max_splitter(data)
#> [1] 143
max_splitter(data, p=0.6)
#> [1] 65