Accumulated Local Effects for ERF

Calculates the Accumulated Local Effects (ALE) from an ERF object

ALE_fn(
  fit,
  var,
  save = TRUE,
  out.folder = NULL,
  cores = parallel::detectCores() - 4
)

Arguments

fit
The fitted object returned from calling ens_random_forests()

var
The name of the response variable

save
A logical flag to save the output as an RData object, default is TRUE.

out.folder
A path to the folder to write out too. If NULL then a folder is generated in the working directory

cores
An integer value that either indicates the number of cores to use for parallel processing or a negative value to indicate the number of cores to leave free. Default is to leave two cores free.

Value

A list that contains the ALEs for each variable, ordered by the mean variable importance

Examples

#run an ERF with 10 RFs and 
ens_rf_ex <- ens_random_forests(df=simData$samples, var="obs", covariates=grep("cov", colnames(simData$samples),value=T), save=FALSE, cores=1)

#> rounding n.forests to the nearest one

ALEdf <- ALE_fn(ens_rf_ex, save=FALSE)

#> No name of response variable, making one

#> rounding n.forests to the nearest one

head(ALEdf[[1]])

#>            x        f.1        f.2        f.3       f.4       f.5        f.6
#> 1 -0.4019663 0.02894524 0.08734524 0.08367872 0.2329606 0.1640701 0.02269326
#> 2 -0.2914307 0.19394524 0.29234524 0.17367872 0.2929606 0.2640701 0.25269326
#> 3 -0.2597696 0.21394524 0.29734524 0.18367872 0.2929606 0.2740701 0.24269326
#> 4 -0.2390396 0.19894524 0.22734524 0.18367872 0.2479606 0.2340701 0.19269326
#> 5 -0.2236097 0.17904474 0.17261887 0.13392747 0.2280601 0.2042194 0.17776789
#> 6 -0.2065132 0.13404474 0.11261887 0.12392747 0.1730601 0.1742194 0.16276789
#>          f.7        f.8         f.9      f.10
#> 1 0.05919643 0.06956431 -0.19536553 0.1470071
#> 2 0.15419643 0.26956431  0.11463447 0.2420071
#> 3 0.15419643 0.28456431  0.11963447 0.2470071
#> 4 0.11919643 0.26456431  0.10463447 0.2270071
#> 5 0.11919643 0.21978819  0.09965935 0.2071066
#> 6 0.10419643 0.19978819  0.08965935 0.1571066