Accumulated Local Effects for ERF

Calculates the Accumulated Local Effects (ALE) from an ERF object

ALE_fn(
  fit,
  var,
  save = TRUE,
  out.folder = NULL,
  cores = parallel::detectCores() - 4
)

Arguments

fit
The fitted object returned from calling ens_random_forests()

var
The name of the response variable

save
A logical flag to save the output as an RData object, default is TRUE.

out.folder
A path to the folder to write out too. If NULL then a folder is generated in the working directory

cores
An integer value that either indicates the number of cores to use for parallel processing or a negative value to indicate the number of cores to leave free. Default is to leave two cores free.

quants
Either a NULL or two numeric values specifying the lower and upper quantile cutoffs to evaluate the ALE for continuous covariates over.

Value

A list that contains the ALEs for each variable, ordered by the mean variable importance

Examples

#run an ERF with 10 RFs and 
ens_rf_ex <- ens_random_forests(df=simData$samples, var="obs", covariates=grep("cov", colnames(simData$samples),value=T), save=FALSE, cores=1)

#> rounding n.forests to the nearest one

ALEdf <- ALE_fn(ens_rf_ex, save=FALSE)

#> No name of response variable, making one

#> rounding n.forests to the nearest one

head(ALEdf[[1]])

#>            x      q        f.1        f.2         f.3        f.4         f.5
#> 1 -0.6321187 0.0002 -0.1144841 -0.1878947 -0.06872968 -0.1200602 -0.04769834
#> 2 -0.5269505 0.0200 -0.2294841 -0.1878947 -0.21372968 -0.2100602 -0.20769834
#> 3 -0.4847148 0.0401 -0.2294841 -0.1878947 -0.21372968 -0.2100602 -0.20769834
#> 4 -0.4507819 0.0601 -0.2194841 -0.1828947 -0.21872968 -0.2000602 -0.20769834
#> 5 -0.4220293 0.0801 -0.2094841 -0.1728947 -0.20872968 -0.1900602 -0.20269834
#> 6 -0.3994806 0.1001 -0.2094841 -0.1728947 -0.22372968 -0.1950602 -0.21269834
#>           f.6        f.7         f.8        f.9        f.10
#> 1 -0.09382639 -0.1180176 -0.05563126 -0.1291141 -0.07723252
#> 2 -0.16382639 -0.2630176 -0.20063126 -0.2141141 -0.20723252
#> 3 -0.16382639 -0.2630176 -0.19565614 -0.2141141 -0.20723252
#> 4 -0.16882639 -0.2530176 -0.19065614 -0.1991141 -0.19223252
#> 5 -0.16382639 -0.2480176 -0.17065614 -0.1941141 -0.17223252
#> 6 -0.16882639 -0.2480176 -0.17065614 -0.2141141 -0.18723252