Ensemble Random Forests

Execute Ensemble Random Forests model on a given dataset

ens_random_forests(
  df,
  var,
  covariates,
  header = NULL,
  out.folder = NULL,
  duplicate = TRUE,
  n.forests = 10L,
  importance = TRUE,
  cores = parallel::detectCores() - 2,
  save = TRUE,
  ntree = 1000,
  mtry = 5,
  var.q = c(0.1, 0.5, 0.9),
  mode = "bin"
)

Arguments

df

A data.frame object

var

A character string indicating the column name of the data frame that contains the number of interactions for the ERF to model; column should be a numeric column

covariates

A character vector indicating the column name(s) of the data frame that contain the covariates

header

A character vector indicating the column name(s) of the data frame that contain the additional columns you wish appended to the output

out.folder

A path to the folder to write out too. If NULL then a folder is generated in the working directory

duplicate

A logical flag that indicates whether to duplicate observations with more than one interaction. Default is TRUE to duplicate all records that interacted with more than one individual (i.e. a fishing set that caught two of the same species)

n.forests

An integer value indicating how many Random Forests to generate in the ensemble, default is 100

importance

A logical flag for the randomForest model to calculate the variable importance

cores

A integer value that either indicates the number of cores to use for parallel processing or a negative value to indicate the number of cores to leave free. Default is to leave two cores free.

save

A logical flag to save the output as an RData object, default is TRUE.

ntree

The number of decision trees to use in each RF, default is 1000

mtry

The number of covariates to try at each node split, default is 5

var.q

The quantiles for the distribution of the variable importance; only exectuted if importance=TRUE

mode

is either 'bin' for binary or if not, then assumed to be multivariate factor, 'bin' is set by default

Value

A list containing the fitted ERF model and associated output.

data: the exact dataset used to fit each Random Forests within the ensemble. At the minimum, it will contain var, covariates, and header if provided. If duplicate==TRUE, then it will also contain the duplicated presence records.

model: the returned fitted individual Random Forests in the ensemble

ens.pred: the ensemble model predictions. This is generally the prediction set to use.

ens.perf: the ensemble model performance metrics including ROC curve.

mu.tr.perf: the mean training set performance across all Random Forests for the AUC, TSS, and RMSE metrics. These are generally useless.

mu.te.perf: the mean test set performance across all Random Forests for the AUC, TSS, and RMSE metrics. These can be informative.

roc_train: the training performance metrics for each Random Forests in the ensemble

roc_test: the test performance metrics for each Random Forests in the ensemble

pred: a list with two objects:

p: predictions from each Random Forests in the ensemble to the dataset

resid: residuals between the observed presence/absences and the predictions from each Random Forests in the ensemble to the dataset

Examples

#run an ERF with 10 RFs and 
ens_rf_ex <- ens_random_forests(df=simData$samples, var="obs",covariates=grep("cov", colnames(simData$samples),value=T), save=FALSE, cores=1)
#> rounding n.forests to the nearest one

# view the dataset used in the model
head(ens_rf_ex$data) 
#>   obs        cov1        cov2       cov3        cov4        cov5      random
#> 1   0 -0.07158000 -0.27811766  0.5741324  0.01366734 -0.15537696 -0.25099787
#> 2   0 -0.12028137  0.26103341 -0.2298234  0.14750795  0.32482174  0.03648563
#> 3   0  0.12422049  0.09052045 -0.3070963  0.14493653  0.02303010 -0.09236283
#> 4   0 -0.12481261 -0.29583389  0.6282519  0.06067430 -0.25240678 -0.54846600
#> 5   0  0.01361460 -0.04989738  0.1600218  0.42189437  0.13166581  0.94890139
#> 6   0  0.01842766  0.18300638 -0.3421439 -0.33128857  0.07998399  2.11019092

#view the model predictions
head(ens_rf_ex$ens.pred) 
#>      P.0    P.1 PRES   resid
#> 1 0.7211 0.2789    0 -0.2789
#> 2 0.3223 0.6777    0 -0.6777
#> 3 0.5530 0.4470    0 -0.4470
#> 4 0.7498 0.2502    0 -0.2502
#> 5 0.3885 0.6115    0 -0.6115
#> 6 0.9884 0.0116    0 -0.0116

#view the mean test threshold-free performance metrics
ens_rf_ex$mu.te.perf 
#>     teAUC    teRMSE     teTSS 
#> 0.7732876 0.3895260 0.5410977 

#view the threshold-free ensemble performance metrics
unlist(ens_rf_ex$ens.perf[c('auc','rmse','tss')]) 
#>       auc      rmse       tss 
#> 0.9721197 0.3797325 0.8184667